feat: Game menu accessibility exploration (#2)
This adds a representative game menu, in both command-line and graphical forms. The goals were: to practice the separation of rendering code, and to create proxy DOM elements for the non-semantic GUI controls.

When using a browser and screen-reader, the menu elements will get announced as the user navigates the menu.